消したファイルと同名だけど大丈夫かな 一応MenuView周りの改築が済んだのでコミット
