+++ NodeAndEdgeView/MenuView.swift
@@ -1,138 +1,144 @@
 //
-//  sideMenuView.swift
+//  NewMenuView.swift
 //  NodeAndEdgeView
 //
-//  Created by Taro on 2019/01/25.
+//  Created by Taro on 2019/02/13.
 //  Copyright © 2019 Taro. All rights reserved.
 //
 
 import UIKit
 
-protocol MenuViewDelegate:NSObjectProtocol{
-    func tappedTextEdit()
-    func tappedCreateEdge()
-    func tappedDeleteNode()
+protocol NewMenuDelegate:NSObjectProtocol {
+    func EditWord()
+    func CreateRelationship()
+    func Remove(removeType:MenuType)
+}
+
+enum MenuType{
+    case Word
+    case Theme
+    case Relationship
 }
 
 class MenuView:UIView{
-    private let itemSet:[String] = ["Text Edit","Create Edge","Delete"]
+    private var editButton = MenuButton()
+    private var relationshipButton = MenuButton()
+    private var removeButton = MenuButton()
     
-    private var defaultHeight:CGFloat = 0.0
-    private var tableView:UITableView = {
-        let table = UITableView()
-            table.backgroundColor = .clear
-            table.tableFooterView = UIView(frame: .zero)
-            table.separatorStyle = .none
-            table.isScrollEnabled = false
-        return table
-    }()
+    private let menuItems:[String] = ["Edit", "Ralationship", "Remove"]
+    private var defaultMenuViewHeight:CGFloat = 0.0
+    private weak var menuDelegate:NewMenuDelegate?
+    private var menuType:MenuType = .Word
     
-    public weak var sideMenuController:MenuViewDelegate?
-    
-    override init(frame: CGRect) {
+    override init(frame:CGRect){
         super.init(frame: frame)
-        self.defaultHeight = frame.height
-        self.setupUI()
-        self.setupItem()
+        self.defaultMenuViewHeight = frame.height
+        self.setView()
+        self.setButton()
+    }
+    
+    public func setDelegate(delegate:NewMenuDelegate){
+        self.menuDelegate = delegate
     }
     
     required init?(coder aDecoder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     
-    private func setupUI(){
+    private func setButton(){
+        self.editButton.frame = CGRect(x: 0,
+                                       y: 0,
+                                       width: self.frame.width,
+                                       height: self.defaultMenuViewHeight/3)
+        self.editButton.setView(name: self.menuItems[0], icon: UIImage(named: "EditIcon")!)
+        self.addSubview(self.editButton)
+        self.editButton.addTarget(self, action: #selector(editButtonDidTapped), for: .touchUpInside)
+        
+        self.relationshipButton.frame = CGRect(x: 0,
+                                               y: self.defaultMenuViewHeight/3,
+                                               width: self.frame.width,
+                                               height: self.defaultMenuViewHeight/3)
+        self.relationshipButton.setView(name: self.menuItems[1], icon: UIImage(named: "RelationshipIcon")!)
+        self.addSubview(self.relationshipButton)
+        self.relationshipButton.addTarget(self, action: #selector(relationshipButtonDidTapped), for: .touchUpInside)
+        
+        self.removeButton.frame = CGRect(x: 0,
+                                         y: self.defaultMenuViewHeight/3 * 2,
+                                         width: self.frame.width,
+                                         height: self.defaultMenuViewHeight/3)
+        self.removeButton.setView(name: self.menuItems[2], icon: UIImage(named: "RemoveIcon")!)
+        self.addSubview(self.removeButton)
+        self.removeButton.addTarget(self, action: #selector(removeButtonDidTapped), for: .touchUpInside)
+    }
+    
+    private func setView(){
         self.backgroundColor = .clear
         self.layer.masksToBounds = true
-        self.layer.cornerRadius = self.frame.width / 8
-        let blurEffectView = UIVisualEffectView(effect: UIBlurEffect(style: .light))
-        blurEffectView.frame = self.bounds
-        self.addSubview(blurEffectView)
+        self.layer.cornerRadius = self.frame.width/8
+        let blur = UIVisualEffectView(effect: UIBlurEffect(style: .light))
+        blur.frame = self.bounds
+        self.addSubview(blur)
     }
     
-    private func setupItem(){
-        self.tableView.frame.size = CGSize(width: self.bounds.width, height: self.bounds.height / 10 * 9)
-        self.tableView.frame.origin = CGPoint(x: self.bounds.origin.x, y: self.bounds.width / 8)
-        self.tableView.rowHeight = (self.bounds.height - self.bounds.width / 8 * 2) / CGFloat(self.itemSet.count)
-        self.tableView.delegate = self
-        self.tableView.dataSource = self
-        self.tableView.register(MenuItemCell.self, forCellReuseIdentifier: "cellID")
-        self.addSubview(tableView)
+    public func changeMenuType(type:MenuType){
+        self.menuType = type
     }
     
-    public func showMenu(isAncestor:Bool){
-        self.frame.size.height = self.defaultHeight
-        if isAncestor{
-            guard let cell = self.tableView.cellForRow(at: IndexPath(row: 2, section: 0)) else {
-                return
-            }
-            cell.isHidden = true
-            self.frame.size.height -= self.tableView.rowHeight
-        }else{
-            guard let cell = self.tableView.cellForRow(at: IndexPath(row: 2, section: 0)) else {
-                return
-            }
-            cell.isHidden = false
+    private func configureMenuView(){
+        switch self.menuType{
+        case .Word:
+            self.editButton.isHidden = false
+            self.relationshipButton.isHidden = false
+            self.removeButton.isHidden = false
+            self.removeButton.frame.origin.y = self.defaultMenuViewHeight/3 * 2
+            self.frame.size.height = self.defaultMenuViewHeight
+        case .Theme:
+            self.editButton.isHidden = false
+            self.relationshipButton.isHidden = false
+            self.removeButton.isHidden = true
+            self.removeButton.frame.origin.y = self.defaultMenuViewHeight/3 * 2
+            self.frame.size.height = self.defaultMenuViewHeight/3 * 2
+        case .Relationship:
+            self.editButton.isHidden = true
+            self.relationshipButton.isHidden = true
+            self.removeButton.isHidden = false
+            self.removeButton.frame.origin.y = 0
+            self.frame.size.height = self.defaultMenuViewHeight/3
         }
-        UIView.animate(withDuration: 0.3, animations: {
+    }
+    
+    public func showMenu(){
+        self.configureMenuView()
+        
+        UIView.animate(withDuration: 0.3){
             self.frame.origin.x -= self.frame.width
-            self.setNeedsLayout()
-        })
+        }
     }
     
     public func hideMenu(){
-        UIView.animate(withDuration: 0.3, animations: {
+        UIView.animate(withDuration: 0.3){
             self.frame.origin.x += self.frame.width
-            if let indexPathForSelectedRow = self.tableView.indexPathForSelectedRow {
-                self.tableView.deselectRow(at: indexPathForSelectedRow, animated: true)
-            }
-            self.setNeedsLayout()
-        })
+        }
     }
     
-    public func unableDelteMode(){
-        guard let cell = self.tableView.cellForRow(at: IndexPath(row: 2, section: 0)) else {
+    @objc func editButtonDidTapped(sender:UIButton){
+        guard let delegate = self.menuDelegate else {
             return
         }
-        cell.isHidden = true
-        self.setNeedsLayout()
-    }
-}
-
-extension MenuView:UITableViewDelegate{
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        if indexPath.row == 0{
-            if let unwrappedMenuController = self.sideMenuController{
-                unwrappedMenuController.tappedTextEdit()
-            }
-        }else if indexPath.row == 1{
-            if let unwrappedSideMenuController = self.sideMenuController{
-                unwrappedSideMenuController.tappedCreateEdge()
-            }
-        }else if indexPath.row == 2{
-            if let unwrappedSideMenuController = self.sideMenuController{
-                unwrappedSideMenuController.tappedDeleteNode()
-            }
-        }else{
-            print("Out of range")
-        }
-    }
-}
-
-extension MenuView:UITableViewDataSource{
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return self.itemSet.count
+        delegate.EditWord()
     }
     
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell:MenuItemCell = tableView.dequeueReusableCell(withIdentifier: "cellID", for: indexPath) as! MenuItemCell
-        if indexPath.row == 0{
-            cell.setImage(name: "TextEditIcon")
-        }else if indexPath.row == 1{
-            cell.setImage(name: "EdgeCreationIcon")
-        }else{
-            cell.setImage(name: "DeleteIcon")
+    @objc func relationshipButtonDidTapped(sender:UIButton){
+        guard let delegate = self.menuDelegate else {
+            return
         }
-        cell.setItem(name: self.itemSet[indexPath.row])
-        return cell
+        delegate.CreateRelationship()
+    }
+    
+    @objc func removeButtonDidTapped(sender:UIButton){
+        guard let delegate = self.menuDelegate else {
+            return
+        }
+        delegate.Remove(removeType: self.menuType)
     }
 }
