+++ NodeAndEdgeView/ViewController.swift
@@ -11,12 +11,8 @@ import UIKit
 class ViewController: UIViewController {
     private let scrollview = ScrollView()
     private var wordViewModel = WordViewModel()
-    private var ralationshipViewModel = RelationshipViewModel()
-    private var menuView = MenuView()
-    
-    // 以下二つのプロパティを改築用に仮に追加(2/13)
-    private var newMenuView = NewMenuView()
-    private var menuType:MenuType = .Word
+    private var relationshipViewModel = RelationshipViewModel()
+    private var newMenuView = MenuView()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -39,15 +35,7 @@ class ViewController: UIViewController {
     }
     
     private func setMenu(){
-        self.menuView = MenuView(frame: CGRect(x: self.view.bounds.width,
-                                               y: self.view.frame.origin.y,
-                                               width: self.view.bounds.width / 4,
-                                               height: self.view.bounds.height / 2))
-        self.menuView.sideMenuController = self
-        self.view.addSubview(self.menuView)
-        
-        // 以下の内容を改築のために仮に設定(左側から出るように設計しているけど，後で右側に変更する)
-        self.newMenuView = NewMenuView(frame: CGRect(x: self.view.bounds.origin.x - self.view.bounds.width/4,
+        self.newMenuView = MenuView(frame: CGRect(x: self.view.bounds.width,
                                                      y: self.view.bounds.origin.y,
                                                      width: self.view.bounds.width / 4,
                                                      height: self.view.bounds.height / 2))
@@ -64,55 +52,10 @@ class ViewController: UIViewController {
         }
         self.scrollview.keepThemeWordViewCenter(wordModel: themeWordModel)
         self.wordSelected(view: self.scrollview, selectedWord: themeWordModel)
-        self.tappedTextEdit()
+        self.EditWord()
     }
 }
 
-
-extension ViewController:MenuViewDelegate{
-    func tappedTextEdit() {
-        let AlertController = UIAlertController(title: "Word", message: "", preferredStyle: .alert)
-        let OKAlertAction = UIAlertAction(title: "OK", style: .default, handler:{[weak AlertController, weak self](action) -> Void in
-            guard let text = AlertController?.textFields?.first?.text else{
-                return
-            }
-            
-            guard let weakself = self else{
-                return
-            }
-            
-            guard let selectedWordModel = weakself.wordViewModel.getSelectedWordModel() else{
-                return
-            }
-            selectedWordModel.setWord(newWord: text)
-            weakself.scrollview.setWordInWordView(wordModel: selectedWordModel, newWord: text)
-            weakself.wordSelected(view: weakself.scrollview, selectedWord: nil)
-        })
-        let CancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: {[weak self](action) -> Void in
-            guard let weakself = self else{
-                return
-            }
-            weakself.wordSelected(view: weakself.scrollview, selectedWord: nil)
-        })
-        AlertController.addTextField{textField in
-            textField.placeholder = "Text"
-            textField.keyboardAppearance  = .dark
-        }
-        AlertController.addAction(OKAlertAction)
-        AlertController.addAction(CancelAction)
-        self.present(AlertController, animated: true, completion: nil)
-    }
-    
-    func tappedDeleteNode() {
-        if let unwrappedSelectedNode = self.wordViewModel.getSelectedWordModel(){
-            self.scrollview.removeWordView(selectedWordModel:unwrappedSelectedNode)
-        }
-    }
-    
-    func tappedCreateEdge() {
-        self.scrollview.changeModeType(mode: .relationshipCreation)
-    }
-}
 
 extension ViewController:WordControlDelegate{
     func createWord(view: ScrollView, position: CGPoint) {
@@ -123,24 +66,14 @@ extension ViewController:WordControlDelegate{
     func wordSelected(view: ScrollView, selectedWord: WordModel?) {
         if let targetUnselectedWord = selectedWord{
             targetUnselectedWord.toggleIsSelected(bool: true)
-            self.menuView.showMenu(isAncestor: self.wordViewModel.getIsThemeWordStatus(targetWordModel: targetUnselectedWord)) // 移行後削除
-            self.wordViewModel.getAllWordModelStatus() // ただのデバッグ用
-            
-            // 改築のために仮配置兼テスト表示用に以下を記述
             self.newMenuView.changeMenuType(type: self.wordViewModel.getIsThemeWordStatus(targetWordModel: targetUnselectedWord) ? .Theme:.Word)
             self.newMenuView.showMenu()
-            
             
             view.changeModeType(mode: .wordSelected)
         }else{
             if let targetSelectedWord = self.wordViewModel.getSelectedWordModel(){
                 targetSelectedWord.toggleIsSelected(bool: false)
-                self.menuView.hideMenu()// 移行後削除
-                self.wordViewModel.getAllWordModelStatus()// ただのデバッグ用
-                
-                // 以下改築のためのテスト表示用に記述
                 self.newMenuView.hideMenu()
-                
                 view.changeModeType(mode: .normal)
                 view.toggleWordViewState(targetWordModel: targetSelectedWord, isSelected: false)
             }
@@ -149,22 +82,17 @@ extension ViewController:WordControlDelegate{
     
     func wordMoved(view: ScrollView, movedWord: WordModel, newPosition: CGPoint) {
         self.wordViewModel.updateWordModelPosition(targetWordModel: movedWord, newPosition: newPosition)
-        view.moveRelationshipView(relationshipModels: self.ralationshipViewModel.searchTargetRelationships(containedWordModel: movedWord))
+        view.moveRelationshipView(relationshipModels: self.relationshipViewModel.getRelatedRelationships(targetWordModel: movedWord))
     }
 
     
     func wordRemoved(view: ScrollView, targetWord: WordModel) {        
         targetWord.toggleIsSelected(bool: false)
-        self.menuView.hideMenu() // 移行後削除
-        let relatedRelationships = self.ralationshipViewModel.searchTargetRelationships(containedWordModel: targetWord)
-        view.deleteRelationshipViews(relationshipModels: relatedRelationships)
-        self.ralationshipViewModel.removeRelationships(relationships: relatedRelationships)
+        let relatedRelationships = self.relationshipViewModel.getRelatedRelationships(targetWordModel: targetWord)
+        self.relationshipRemoved(view: view, targetRelationships: relatedRelationships)
+        view.removeWordView(selectedWordModel: targetWord)
         self.wordViewModel.removeWordModel(targetWordModel: targetWord)
         view.changeModeType(mode: .normal)
-        self.newMenuView.hideMenu()
-        
-        self.wordViewModel.getAllWordModelStatus()// ただのデバッグ用
-        self.ralationshipViewModel.getAllRelationships()// ただのデバッグ用
     }
 }
 
@@ -177,32 +105,34 @@ extension ViewController:RelationshipControlDelegate{
     }
     
     func createRelationship(view: ScrollView, dst: WordModel) {
-        if let unwrappedParentNode = self.wordViewModel.getSelectedWordModel(){
-            let newEdge = RelationshipModel(scrWordModel: unwrappedParentNode, dstWordModel: dst)
-            if self.ralationshipViewModel.addRelationship(newRelationship: newEdge){
-                self.ralationshipViewModel.getAllRelationships()
-                view.createRelationshipView(src: unwrappedParentNode, dst: dst, newRelationshipModel: newEdge)
+        if let srcWordModel = self.wordViewModel.getSelectedWordModel(){
+            let newRelationshipModel = RelationshipModel(scrWordModel: srcWordModel, dstWordModel: dst)
+            if self.relationshipViewModel.addRelationship(newRelationship: newRelationshipModel){
+                view.createRelationshipView(src: srcWordModel, dst: dst, newRelationshipModel: newRelationshipModel)
             }
         }else{
             print("Edge creation failed")
         }
     }
     
-    func relationshipRemoved(view: ScrollView, targetRelationship: RelationshipModel) {
-        let deleteEdgeModel:[RelationshipModel] = [targetRelationship]
-        self.ralationshipViewModel.removeRelationships(relationships: deleteEdgeModel)
-        view.deleteRelationshipViews(relationshipModels: deleteEdgeModel)
-        self.wordSelected(view: self.scrollview, selectedWord: nil)
+    func relationshipRemoved(view: ScrollView, targetRelationships: [RelationshipModel]) {
+        view.removeRelationshipViews(relationshipModels: targetRelationships)
+        view.changeModeType(mode: .normal)
+        self.relationshipViewModel.removeRelationships(relationships: targetRelationships)
     }
     
     func relationshipSelected(view: ScrollView, targetRelationship: RelationshipModel?) {
-        if let _ = targetRelationship{
+        if let targetRelationship = targetRelationship{
+            targetRelationship.toggleIsSelected(bool: true)
             view.changeModeType(mode: .ralationshipSelected)
             self.newMenuView.changeMenuType(type: .Relationship)
             self.newMenuView.showMenu()
         }else{
-            self.newMenuView.hideMenu()
-            view.changeModeType(mode: .normal)
+            if let selectedRelationshipModel = self.relationshipViewModel.getSelectedRelationshipModel(){
+                selectedRelationshipModel.toggleIsSelected(bool: false)
+                self.newMenuView.hideMenu()
+                view.changeModeType(mode: .normal)
+            }
         }
     }
 }
@@ -245,12 +175,21 @@ extension ViewController:NewMenuDelegate{
         self.scrollview.changeModeType(mode: .relationshipCreation)
     }
     
-    func Remove() {
-        // TODO
-        // 消すのはWordModelに限らないので，場合分けが必要
-        // というか．Relationshipに削除とか選択って概念がないから．そこから作らないといけない
-        if let selectedWordModel = self.wordViewModel.getSelectedWordModel(){
-            self.scrollview.removeWordView(selectedWordModel:selectedWordModel)
+    func Remove(removeType:MenuType) {
+        switch removeType{
+        case .Word:
+            if let selectedWordModel = self.wordViewModel.getSelectedWordModel(){
+                self.wordRemoved(view: self.scrollview, targetWord: selectedWordModel)
+            }
+        case .Relationship:
+            if let selectedRelationshipModel = self.relationshipViewModel.getSelectedRelationshipModel(){
+                let removeRelationshipList:[RelationshipModel] = [selectedRelationshipModel]
+                self.relationshipRemoved(view: self.scrollview, targetRelationships: removeRelationshipList)
+            }
+        case .Theme:
+            print("ERROR: Theme word cannot be removed")
+            return
         }
+        self.newMenuView.hideMenu()
     }
 }
